Move progress bar configuration to separate class
By moving the progress bar configuration into a separate class
we can make the run command more lean. This will allow some improvements
in the future.
