+++ packages/guides-cli/src/Command/Run.php
@@ -13,6 +13,7 @@ declare(strict_types=1);
 
 namespace phpDocumentor\Guides\Cli\Command;
 
+use Doctrine\Deprecations\Deprecation;
 use Flyfinder\Path;
 use Flyfinder\Specification\InPath;
 use Flyfinder\Specification\NotSpecification;
@@ -25,15 +26,7 @@ use Monolog\Logger;
 use phpDocumentor\FileSystem\FlySystemAdapter;
 use phpDocumentor\Guides\Cli\Logger\SpyProcessor;
 use phpDocumentor\Guides\Compiler\CompilerContext;
-use phpDocumentor\Guides\Event\PostCollectFilesForParsingEvent;
-use phpDocumentor\Guides\Event\PostParseDocument;
-use phpDocumentor\Guides\Event\PostParseProcess;
 use phpDocumentor\Guides\Event\PostProjectNodeCreated;
-use phpDocumentor\Guides\Event\PostRenderDocument;
-use phpDocumentor\Guides\Event\PostRenderProcess;
-use phpDocumentor\Guides\Event\PreParseDocument;
-use phpDocumentor\Guides\Event\PreRenderDocument;
-use phpDocumentor\Guides\Event\PreRenderProcess;
 use phpDocumentor\Guides\Handlers\CompileDocumentsCommand;
 use phpDocumentor\Guides\Handlers\ParseDirectoryCommand;
 use phpDocumentor\Guides\Handlers\ParseFileCommand;
@@ -46,7 +39,6 @@ use Psr\Clock\ClockInterface;
 use Psr\Log\LogLevel;
 use RuntimeException;
 use Symfony\Component\Console\Command\Command;
-use Symfony\Component\Console\Helper\ProgressBar;
 use Symfony\Component\Console\Input\InputArgument;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Input\InputOption;
@@ -63,7 +55,6 @@ use function count;
 use function implode;
 use function is_countable;
 use function is_dir;
-use function microtime;
 use function pathinfo;
 use function sprintf;
 use function strtoupper;
@@ -77,6 +68,7 @@ final class Run extends Command
         private readonly SettingsManager $settingsManager,
         private readonly ClockInterface $clock,
         private readonly EventDispatcher $eventDispatcher,
+        private readonly ProgressBarSubscriber $progressBarSubscriber,
     ) {
         parent::__construct('run');
 
@@ -156,76 +148,12 @@ final class Run extends Command
 
     public function registerProgressBar(ConsoleOutputInterface $output): void
     {
-        $parsingProgressBar = new ProgressBar($output->section());
-        $parsingProgressBar->setFormat('Parsing: %current%/%max% [%bar%] %percent:3s%% %message%');
-        $parsingStartTime = microtime(true);
-        $this->eventDispatcher->addListener(
-            PostCollectFilesForParsingEvent::class,
-            static function (PostCollectFilesForParsingEvent $event) use ($parsingProgressBar, &$parsingStartTime): void {
-                // Each File needs to be first parsed then rendered
-                $parsingStartTime = microtime(true);
-                $parsingProgressBar->setMaxSteps(count($event->getFiles()));
-            },
+        Deprecation::trigger(
+            'phpdocumentor/guides-cli',
+            'https://github.com/phpdocumentor/guides/issues/33',
+            'Progressbar will be registered via settings',
         );
-        $this->eventDispatcher->addListener(
-            PreParseDocument::class,
-            static function (PreParseDocument $event) use ($parsingProgressBar): void {
-                $parsingProgressBar->setMessage('Parsing file: ' . $event->getFileName());
-                $parsingProgressBar->display();
-            },
-        );
-        $this->eventDispatcher->addListener(
-            PostParseDocument::class,
-            static function (PostParseDocument $event) use ($parsingProgressBar): void {
-                $parsingProgressBar->advance();
-            },
-        );
-        $this->eventDispatcher->addListener(
-            PostParseProcess::class,
-            static function (PostParseProcess $event) use ($parsingProgressBar, $parsingStartTime): void {
-                $parsingTimeElapsed = microtime(true) - $parsingStartTime;
-                $parsingProgressBar->setMessage(sprintf(
-                    'Parsed %s files in %.2f seconds',
-                    $parsingProgressBar->getMaxSteps(),
-                    $parsingTimeElapsed,
-                ));
-                $parsingProgressBar->finish();
-            },
-        );
-        $that = $this;
-        $this->eventDispatcher->addListener(
-            PreRenderProcess::class,
-            static function (PreRenderProcess $event) use ($that, $output): void {
-                $renderingProgressBar = new ProgressBar($output->section(), count($event->getCommand()->getDocumentArray()));
-                $renderingProgressBar->setFormat('Rendering: %current%/%max% [%bar%] %percent:3s%% Output format ' . $event->getCommand()->getOutputFormat() . ': %message%');
-                $renderingStartTime = microtime(true);
-                $that->eventDispatcher->addListener(
-                    PreRenderDocument::class,
-                    static function (PreRenderDocument $event) use ($renderingProgressBar): void {
-                        $renderingProgressBar->setMessage('Rendering: ' . $event->getCommand()->getFileDestination());
-                        $renderingProgressBar->display();
-                    },
-                );
-                $that->eventDispatcher->addListener(
-                    PostRenderDocument::class,
-                    static function (PostRenderDocument $event) use ($renderingProgressBar): void {
-                        $renderingProgressBar->advance();
-                    },
-                );
-                $that->eventDispatcher->addListener(
-                    PostRenderProcess::class,
-                    static function (PostRenderProcess $event) use ($renderingProgressBar, $renderingStartTime): void {
-                        $renderingElapsedTime = microtime(true) - $renderingStartTime;
-                        $renderingProgressBar->setMessage(sprintf(
-                            'Rendered %s documents in %.2f seconds',
-                            $renderingProgressBar->getMaxSteps(),
-                            $renderingElapsedTime,
-                        ));
-                        $renderingProgressBar->finish();
-                    },
-                );
-            },
-        );
+        $this->progressBarSubscriber->subscribe($output, $this->eventDispatcher);
     }
 
     private function getSettingsOverriddenWithInput(InputInterface $input): ProjectSettings
@@ -323,7 +251,7 @@ final class Run extends Command
 
 
         if ($output instanceof ConsoleOutputInterface && $settings->isShowProgressBar()) {
-            $this->registerProgressBar($output);
+            $this->progressBarSubscriber->subscribe($output, $this->eventDispatcher);
         }
 
         if ($settings->getInputFile() === '') {
